format_message() uses REPLACE_MAP, kwargs
This commit changes the way messages are composed internally. format_message()
(formerly format_np()) takes a template and any number of keyword
arguments. It then replaces all placeholders (such as [artist]) as long
as they are linked to an available kwarg through the REPLACE_MAP. This
way it is much easier to add new placeholders.