Handling negative API response
In lastfmnp and tellnp there were bugs causing the script to post the
raw message template to the buffer (in the case of tellnp) or failing
and printing a stacktrace (in the case of lastfmnp) whenever the API
responded that there is no "now playing"-information available. This
patch handles this case correctly by not trying to print an empty
message and instead adding a short error message on the core buffer.