_match_list_or_word()
Does as the name suggests. The result is always a list, in the single
word case the result is a list with one element. The input list looks
like this: (anything, in, here, except, more parenthesis). Nested lists
are not allowed and will not be parsed! That means (a, (b, c)) will not
return a list with two elements a and (b, c) but simply refuse to parse.