Subshells and environment
lfm now knows a subshell $ invar { command | chain } outvar; which is an
entire command chain nested. You can specify invars which will be fed as
input to the first command in the command chain and an outvar which will
store the output of the last command in the chain. invar and outvar are
optional.