Check for weechat::register()
This final patch completes the series to add an abstraction layer around
weechat's config support. The script can be used as standalone perl
script by setting LFM_* environment variables. They follow the same
naming pattern as weechat options.

To use this script in another environment, it is likely that you have to
adapt the following:
* lfm_print() lfm_error() lfm_info()
* cnf() and cnf_str()
* cnf_set_default()
* alias_path()
* the loading/starting routine on the bottom of the script